Finished.
All the functionalities has been finished.
The program allows the user to add contacts, with name and phone, and remove them. If the user tries to add an existing contact, the program shows an alert and doesn't add it. Also, if the user tries to remove an unexisting contact, the program do the same.
Added to that, the program shows the existing contacts and updates the information when the user adds or removes something.